+++ server/src/services/memories/helpersService.ts
@@ -1,108 +1,10 @@
-import { pipeline } from 'node:stream/promises';
-import { Readable } from 'node:stream';
-import { Response } from 'express';
 import { canAccessTrip, db } from "../../db/database";
-import { safeFetch, SsrfBlockedError, type SafeFetchOptions } from '../../utils/ssrfGuard';
 import { decrypt_api_key } from '../../nest/common/crypto/apiKeyCrypto';
-
-// helpers for handling return types
-
-type ServiceError = { success: false; error: { message: string; status: number } };
-export type ServiceResult<T> = { success: true; data: T } | ServiceError;
-
-
-export function fail(error: string, status: number): ServiceError {
-    return { success: false, error: { message: error, status } };
-}
-
-
-export function success<T>(data: T): ServiceResult<T> {
-    return { success: true, data: data };
-}
-
-
-export function mapDbError(error: Error, fallbackMessage: string): ServiceError {
-    if (error && /unique|constraint/i.test(error.message)) {
-        return fail('Resource already exists', 409);
-    }
-    return fail(error.message, 500);
-}
-
-
-export function handleServiceResult<T>(res: Response, result: ServiceResult<T>): void {
-    if ('error' in result) {
-        res.status(result.error.status).json({ error: result.error.message });
-    }
-    else {
-        res.json(result.data);
-    }
-}
-
-// ----------------------------------------------
-// types used across memories services
-export type Selection = {
-    provider: string;
-    asset_ids: string[];
-    passphrase?: string;
-};
-
-export type StatusResult = {
-    connected: true;
-    user: { name: string }
-} | {
-    connected: false;
-    error: string
-};
-
-export type SyncAlbumResult = {
-    added: number;
-    total: number
-};
-
-
-export type AlbumsList = {
-    albums: Array<{ id: string; albumName: string; assetCount: number; passphrase?: string }>
-};
-
-export type Asset = {
-    id: string;
-    takenAt: string;
-    mediaType?: string;
-    city?: string | null;
-    country?: string | null;
-    lat?: number | null;
-    lng?: number | null;
-};
-
-export type AssetsList = {
-    assets: Asset[],
-    total: number,
-    hasMore: boolean
-};
-
-
-export type AssetInfo = {
-    id: string;
-    takenAt: string | null;
-    city: string | null;
-    country: string | null;
-    state?: string | null;
-    camera?: string | null;
-    lens?: string | null;
-    focalLength?: string | number | null;
-    aperture?: string | number | null;
-    shutter?: string | number | null;
-    iso?: string | number | null;
-    lat?: number | null;
-    lng?: number | null;
-    orientation?: number | null;
-    description?: string | null;
-    width?: number | null;
-    height?: number | null;
-    fileSize?: number | null;
-    fileName?: string | null;
-}
-
+// The ServiceResult envelope, the asset shapes and pipeAsset moved to
+// nest/memories/memories.helpers.ts — this file is down to the DB-backed
+// access checks and album-link lookups, and folds with the providers.
+import { fail, success, type ServiceResult } from '../../nest/memories/memories.helpers';
+export * from '../../nest/memories/memories.helpers';
 
 //for loading routes to settings page, and validating which services user has connected
 type PhotoProviderConfig = {
@@ -254,44 +156,4 @@ export function getAlbumLinkForSync(tripId: string, linkId: string, userId: numb
 
 export function updateSyncTimeForAlbumLink(linkId: string): void {
     db.prepare('UPDATE trip_album_links SET last_synced_at = CURRENT_TIMESTAMP WHERE id = ?').run(linkId);
-}
-
-export async function pipeAsset(url: string, response: Response, headers?: Record<string, string>, signal?: AbortSignal, defaultCacheControl?: string, fetchOptions?: SafeFetchOptions): Promise<void> {
-    try {
-        const resp = await safeFetch(url, { headers, signal: signal as any }, fetchOptions);
-
-        response.status(resp.status);
-        if (resp.headers.get('content-type')) response.set('Content-Type', resp.headers.get('content-type') as string);
-        if (!resp.ok) {
-            response.set('Cache-Control', 'no-store, max-age=0');
-        } else if (resp.headers.get('cache-control')) {
-            response.set('Cache-Control', resp.headers.get('cache-control') as string);
-        } else if (defaultCacheControl) {
-            response.set('Cache-Control', defaultCacheControl);
-        }
-        if (resp.headers.get('content-length')) response.set('Content-Length', resp.headers.get('content-length') as string);
-        if (resp.headers.get('content-disposition')) response.set('Content-Disposition', resp.headers.get('content-disposition') as string);
-        // Pass byte-range metadata through so a <video> can seek (#823). Upstream
-        // returns 206 + Content-Range when the caller forwarded a Range header.
-        if (resp.headers.get('accept-ranges')) response.set('Accept-Ranges', resp.headers.get('accept-ranges') as string);
-        if (resp.headers.get('content-range')) response.set('Content-Range', resp.headers.get('content-range') as string);
-
-        if (!resp.body) {
-            response.end();
-        } else {
-            await pipeline(Readable.fromWeb(resp.body as any), response);
-        }
-    } catch (error) {
-        if (response.headersSent) {
-            response.end();
-            return;
-        }
-        if (error instanceof SsrfBlockedError) {
-            response.status(400).json({ error: error.message });
-        } else {
-            // Don't log the URL — it can carry a Synology _sid / passphrase.
-            console.error('pipeAsset: upstream fetch failed:', error);
-            response.status(500).json({ error: 'Failed to fetch asset' });
-        }
-    }
 }
